move getdata.c and color2rgb.c and headers to the Source shared directory since these files are used by more than one program - make corresponding changes Makefiles
