+++ Source/shared/getdata.c
