+++ Source/shared/getdata.h
